implement property add and get api

+++ src/app/module/User/User.services.ts
@@ -1,6 +1,5 @@
 import { Prisma, UserStatus } from "@prisma/client";
 import { sortOrderType } from "../../constants/common";
-import ApiError from "../../error/ApiError";
 import fieldValidityChecker from "../../utils/fieldValidityChecker";
 import pagination from "../../utils/pagination";
 import {
@@ -11,8 +10,10 @@ import {
 import prisma from "../../shared/prisma";
 import { TAuthUser } from "../../interfaces/common";
 import { TFile } from "../../interfaces/file";
-import { fileUploader } from "../../utils/fileUploader";
 import { TImage } from "../Image/Image.interfaces";
+import supabase from "../../shared/supabase";
+import config from "../../config";
+import ApiError from "../../error/ApiError";
 
 const getUsers = async (query: Record<string, any>) => {
   const { searchTerm, page, limit, sortBy, sortOrder, id, ...remainingQuery } =
@@ -105,7 +106,7 @@ const updateProfile = async (
   payload: Record<string, any>,
   file: TFile | undefined
 ) => {
-  const image: Record<string, any> = {};
+  let image: Record<string, any> = {};
   if (file) {
     const userInfo = await prisma.user.findUniqueOrThrow({
       where: {
@@ -123,21 +124,38 @@ const updateProfile = async (
         },
       });
       if (profile_pic_info) {
-        await fileUploader.deleteToCloudinary([profile_pic_info.cloud_id]);
-        await prisma.image.delete({
-          where: {
-            id: profile_pic_info.id,
-          },
-        });
+        const { data, error } = await supabase.storage
+          .from(config.supabase_bucket_name)
+          .remove([profile_pic_info.path]);
+        if (data) {
+          await prisma.image.delete({
+            where: {
+              id: profile_pic_info.id,
+            },
+          });
+        }
       }
     }
 
-    const convertedFile = Buffer.from(file.buffer).toString("base64");
-    const dataURI = `data:${file.mimetype};base64,${convertedFile}`;
-    const cludinaryResponse = await fileUploader.uploadToCloudinary(dataURI);
-    image["path"] = cludinaryResponse?.secure_url as string;
-    image["cloud_id"] = cludinaryResponse?.public_id as string;
-    image["name"] = file.originalname;
+    const fileName = `${Date.now()}_${file.originalname}`;
+    const { data } = await supabase.storage
+      .from(config.supabase_bucket_name)
+      .upload(fileName, file.buffer, {
+        contentType: file.mimetype,
+      });
+
+    if (!data?.id) {
+      throw new ApiError(
+        httpStatus.INTERNAL_SERVER_ERROR,
+        "Failed to upload profile picture"
+      );
+    }
+
+    image = {
+      name: file.originalname,
+      path: data.path,
+      bucket_id: data.id,
+    };
   }
 
   let profilePic;
